backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
# imaplib ist Teil der Standardbibliothek 